Reorganização das classes de configuração
