+++ backend/agricultura-familiar/src/main/java/com/es/agriculturafamiliar/config/database/RetryDatabaseConfiguration.java
@@ -1,4 +1,4 @@
-package com.es.agriculturafamiliar.config;
+package com.es.agriculturafamiliar.config.database;
 
 import javax.sql.DataSource;
