+++ backend/agricultura-familiar/src/main/java/com/es/agriculturafamiliar/config/email/AsyncMessageSenderFactory.java
@@ -1,4 +1,4 @@
-package com.es.agriculturafamiliar.config;
+package com.es.agriculturafamiliar.config.email;
 
 public abstract class AsyncMessageSenderFactory<T, U, K> {
     public abstract K createAsyncMessageSender(T data, U sender);
