+++ backend/agricultura-familiar/src/main/java/com/es/agriculturafamiliar/config/email/EmailProperties.java
@@ -1,4 +1,4 @@
-package com.es.agriculturafamiliar.config;
+package com.es.agriculturafamiliar.config.email;
 
 import org.springframework.context.annotation.Configuration;
 import lombok.Getter;
