+++ backend/agricultura-familiar/src/main/java/com/es/agriculturafamiliar/config/email/EmailSenderServiceFactory.java
@@ -1,4 +1,4 @@
-package com.es.agriculturafamiliar.config;
+package com.es.agriculturafamiliar.config.email;
 
 import java.nio.charset.StandardCharsets;
 import java.util.Map;
