+++ backend/agricultura-familiar/src/main/java/com/es/agriculturafamiliar/config/security/AuthenticationValidatorConfig.java
@@ -1,4 +1,4 @@
-package com.es.agriculturafamiliar.config;
+package com.es.agriculturafamiliar.config.security;
 
 import java.util.List;
