+++ backend/agricultura-familiar/src/main/java/com/es/agriculturafamiliar/config/security/JwtFilter.java
@@ -1,4 +1,4 @@
-package com.es.agriculturafamiliar.config;
+package com.es.agriculturafamiliar.config.security;
 
 import java.io.IOException;
