+++ backend/agricultura-familiar/src/main/java/com/es/agriculturafamiliar/config/security/PasswordEncoderConfiguration.java
@@ -1,4 +1,4 @@
-package com.es.agriculturafamiliar.config;
+package com.es.agriculturafamiliar.config.security;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
