+++ backend/agricultura-familiar/src/main/java/com/es/agriculturafamiliar/config/security/WebSecurityConfig.java
@@ -1,4 +1,4 @@
-package com.es.agriculturafamiliar.config;
+package com.es.agriculturafamiliar.config.security;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
