+++ backend/agricultura-familiar/src/main/java/com/es/agriculturafamiliar/config/web/FrontEndProperties.java
@@ -1,4 +1,4 @@
-package com.es.agriculturafamiliar.config;
+package com.es.agriculturafamiliar.config.web;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
