+++ backend/agricultura-familiar/src/main/java/com/es/agriculturafamiliar/listener/RenovacaoCodigoConfirmacaoListener.java
@@ -6,8 +6,8 @@ import org.springframework.context.event.EventListener;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Component;
 
-import com.es.agriculturafamiliar.config.EmailProperties;
-import com.es.agriculturafamiliar.config.FrontEndProperties;
+import com.es.agriculturafamiliar.config.email.EmailProperties;
+import com.es.agriculturafamiliar.config.web.FrontEndProperties;
 import com.es.agriculturafamiliar.entity.Email;
 import com.es.agriculturafamiliar.enums.TemplateType;
 import com.es.agriculturafamiliar.event.RenovacaoCodigoConfirmacaoEvent;
