+++ backend/agricultura-familiar/src/main/resources/application.properties
@@ -6,7 +6,7 @@ server.servlet.context-path=/agricultura-familiar/api
 security.basic.enabled=true
 
 #Active profile
-spring.profiles.active=dev
+spring.profiles.active=docker
 
 
 # JWT config
